Add platformio project and source to test the BMP, and two post-apogee detection methods
There are two loop methods implemented. The first to test the BMP and the latter with a state machine with ASCENT->DESCENT->RECOVERY transitions.
ASCENT->DESCENT transition is based on one of two methods (Altitude delta and Last K altitudes average delta respectively)
DESCENT->RECOVERY transition is based on the activation of the parachute.